+++ template/src/config.ts
@@ -19,3 +19,8 @@ export const DEPLOYMENT_ENV = (
 
 // This is the actual complete public URL during runtime of the deployed frontend app.
 export const PUBLIC_URL = window.location.origin + BASE_NAME;
+
+export const LOADING_INDICATOR_DELAY_MS = 300;
+
+// For debugging only: Make /dashboard route publicly accessible
+export const DEBUG_PUBLIC_DASHBOARD = false;