+++ CHANGELOG.md
@@ -10,6 +10,14 @@
 - Use `craco` instead of `react-scripts` so that the `craco.config.js` file is used and the license checker works
 - Remove service worker file
 - Remove "REACT_APP_DEPLOYMENT_ENV" variable
+- Remove @types/yup
+- Add loading indicator delay
+- Add useQuery
+- Add useHash
+- Add useInterval
+- Add TextStyles.css
+- Renamed Styles.ts to Colors.ts
+
 
 ## 2.1.0